Move test_binops
